requirements.txt
Django==1.9.7
django-selectable==0.9.0
